requirements-test.txt
mock>=1.0.1
pytest-mock
pytest
pytest-django
pytest-cov

# Additional test requirements go here
